Fixed a crash, added stuff in master nexus.
Crash was caused by not reading a variable in the lowest call stack. Those variables are global, or belong to the owner, and are therefore valid to be called inside a function.

Added a sorting algorithams, binary search, visual representation of vector additions, and examples of operator overloads and implicit conversions.